refactor base classes
uses long rather than int as the key field.
allows setting of prefix and suffix
allows setting of key field name
makes all methods virtual

+++ Dapper.Rainbow/SqlCompactDatabase.cs
@@ -19,7 +19,7 @@ namespace Dapper.Rainbow
             /// </summary>
             /// <param name="data">Either DynamicParameters or an anonymous type or concrete type</param>
             /// <returns></returns>
-            public override int? Insert(dynamic data)
+            public override long? Insert(dynamic data)
             {
                 var o = (object)data;
                 List<string> paramNames = GetParamNames(o);
